Fuse variables to a recipient dataset using CART fusion model

Fuse variables to a recipient dataset using a fusion model object produced by train(). fuseM() provides a convenience wrapper for generating multiple implicates.

Usage

fuseCART(data, train.object, ignore_self = FALSE)

Arguments

data

Data frame. Recipient dataset. All categorical variables should be factors and ordered whenever possible. Data types and levels are strictly validated against predictor variables defined in train.object.

train.object

Output from a successful call to train.

ignore_self

Logical. If TRUE, the simulation step excludes "self matches" (i.e. row 1 in data cannot match with row 1 in the original donor. Only useful for validation exercises. Do not use otherwise.

Value

For fuse(), a data frame with same number of rows as data and one column for each synthetic fusion variable. The order of the columns reflects the order in which they where fused.

For fuseM(), a data frame with number of rows equal to M * nrow(data). Integer column ".M" indicates implicate assignment of each observation. Note that the ordering of recipient observations is consistent within implicates, so do not change the row order if using with analyze().

Examples

# Build a fusion model using RECS microdata
?recs
fusion.vars <- c("electricity", "natural_gas", "aircon")
predictor.vars <- names(recs)[2:12]
fit <- train(data = recs, y = fusion.vars, x = predictor.vars)
#> 3 fusion variables
#> 11 initial predictor variables
#> 5686 observations
#> Using all available predictors for each fusion variable
#> Training step 1 of 3: electricity
#> -- R-squared of cluster means: 0.973 
#> -- Number of neighbors in each cluster:
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>    10.0    70.0   148.0   192.5   296.2   498.0 
#> Training step 2 of 3: natural_gas
#> -- R-squared of cluster means: 0.976 
#> -- Number of neighbors in each cluster:
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>    10.0    62.0   146.0   192.2   307.5   498.0 
#> Training step 3 of 3: aircon
#> Fusion model saved to:
#>  /home/runner/work/fusionModel/fusionModel/docs/reference/fusion_model.fsn 
#> Total processing time: 6.97 secs 

# Generate single implicate of synthetic 'fusion.vars',
#  using original RECS data as the recipient
recipient <- recs[predictor.vars]
sim <- fuse(data = recipient, train.object = fit)
#> Error in fuse(data = recipient, train.object = fit): unused argument (train.object = fit)
head(sim)
#> Error in head(sim): object 'sim' not found

# Generate multiple implicates
sim <- fuseM(data = recipient, train.object = fit, M = 5)
#> Error in fuseM(data = recipient, train.object = fit, M = 5): could not find function "fuseM"
head(sim)
#> Error in head(sim): object 'sim' not found
table(sim$.M)
#> Error in table(sim$.M): object 'sim' not found