Read ORNL UrbanPop data from disk

NOTE: For fusionACS internal use only! Efficiently read pre-processed ORNL UrbanPop data from disk. Function arguments only make sense if you are familiar with the structure of the processed UrbanPop data.

Usage

read_up(
  path,
  year = NULL,
  state = NULL,
  county = NULL,
  tract_bg = NULL,
  hid = NULL,
  df = NULL,
  cores = max(1, parallel::detectCores(logical = FALSE) - 1)
)

Arguments

path

Character. File path to fst file containing UrbanPop data.

year

Integer. Year(s) to select.

state

Integer. State FIPS code(s) to select.

county

Integer. County FIPS code(s) to select.

tract_bg

Integer. Tract and block group FIPS code(s) to select.

hid

Integer. ACS-PUMS household ID(s) to select.

df

Data frame containing at least 'year' and/or 'state' columns. Provides unique combinations of the above argument values to return. df is used to perform an inner merge on an initial subset of data based on 'year' or 'state'.

cores

Integer. Number of cores used by read_fst.

Value

A keyed data.table.

Details

Provides an efficient and fast way to load a subset of UrbanPop data into memory. An initial subset is read using state or year to restrict rows. Then data.table operations (subset or merge) are used to efficiently reduce to the final returned subset.

Examples

up.path <- "~/Documents/Projects/fusionData/urbanpop/Processed national UrbanPop.fst"

out <- read_up(path = up.path, year = 2015, state = 4)
unique(dplyr::select(out, year, state))

out <- read_up(path = up.path, year = 2015:2016, state = c(2, 12, 15))
unique(dplyr::select(out, year, state))

up.df <- data.frame(year = c(2015, 2018), state = c(8, 5), county = c(1, 3))
out <- read_up(path = up.path, df = up.df)
unique(dplyr::select(out, year, state, county))