Train a fusion model using Classification and Regression Trees (CART)

Train a fusion model on donor data using CART implemented via rpart.

Usage

trainCART(
  data,
  y,
  x,
  weight = NULL,
  order = NULL,
  deriv = TRUE,
  smoothed = FALSE,
  cores = 1,
  lasso = NULL,
  maxcats = 12,
  complexity = 0,
  cvfolds = 0,
  cvfactor = 0,
  node.obs = c(10, 10),
  initial = c(1, 0)
)

Arguments

data

Data frame. Donor dataset. Categorical variables must be factors and ordered whenever possible.

y

Character vector. The variables to fuse to a recipient dataset.

x

Character vector. Predictor variables common to donor and eventual recipient.

weight

Character vector. Name of the observation weights column. If NULL (default), uniform weights are assumed.

order

Character vector. Order in which to fuse y. If NULL (default), a pseudo-optimal order is determined internally.

deriv

Logical. Should algorithm check for derivative relationships prior to building fusion models?

smoothed

Logical. Should the synthetic values be smoothed via KDE? Default is FALSE.

cores

Integer. Number of cores used for parallel operations. Passed to cl argument of pblapply. Ignored on Windows systems.

lasso

Numeric (0-1) or NULL. Controls extent of predictor variable pre-screening via LASSO regression. If NULL (default), no screening is performed. lasso = 1 invokes the least-restrictive screening. See Details.

maxcats

Positive integer or NULL. Maximum number of levels allowed in an unordered factor predictor variable when the response (fusion) variable is also an unordered factor. Prevents excessive rpart computation time. A K-means clustering strategy is used to cluster the predictor to no more than maxcats levels.

complexity

Numeric. Passed to cp argument of rpart.control to control complexity of decision trees.

cvfolds

Integer. Number of cross-validation folds used by rpart to determine optimal tree complexity. Default is no cross-validation (cvfolds = 0).

cvfactor

Numeric. Controls how decision trees are pruned when cvfolds > 0. cvfactor = 0 (the default) selects the tree complexity that minimizes the cross-validation error. cvfactor = 1 is equivalent to Breiman's "1-SE" rule.

node.obs

Numeric vector of length 2. Minimum number of observations in tree nodes. First number is for numeric response variables; second for categorical. Each is passed to minbucket argument of rpart.control.

initial

Numeric vector of length 2. Controls speed/performance of the initial model-fitting routune used to determine fusion order. First number is proportion of data observations to randomly sample. Second number is the cp argument passed to rpart.control. See Details.

Value

A list containing trained model information to be passed to fuse.

Details

When lasso is non-NULL, predictor variables are "pre-screened" using LASSO regression via glmnet prior to fitting a rpart model. Predictors with a LASSO coefficient of zero are excluded from consideration. This can speed up tree-fitting considerably when data is large. Lower values of lasso are more aggressive at excluding predictors; the LASSO lambda is chosen such that the deviance explained is at least lasso-% of the maximum. To ensure the LASSO step is fast, pre-screening is only used for numeric, logical, and ordered factor response variables (the latter integerized).

Since determination of the fusion order only makes use of variable importance results from the initial (fully-specified) models, employing a random subset of observations and less complex models (controlled via initial) can yield a competitive fusion order at less expense.

Examples

# Build a fusion model using RECS microdata
?recs
fusion.vars <- c("electricity", "natural_gas", "aircon")
predictor.vars <- names(recs)[2:12]
fit <- train(data = recs, y = fusion.vars, x = predictor.vars)
#> 3 fusion variables
#> 11 initial predictor variables
#> 5686 observations
#> Using all available predictors for each fusion variable
#> Training step 1 of 3: electricity
#> -- R-squared of cluster means: 0.972 
#> -- Number of neighbors in each cluster:
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>    10.0    72.0   152.5   195.6   300.0   498.0 
#> Training step 2 of 3: natural_gas
#> -- R-squared of cluster means: 0.976 
#> -- Number of neighbors in each cluster:
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>    10.0    62.0   147.0   191.9   312.0   498.0 
#> Training step 3 of 3: aircon
#> Fusion model saved to:
#>  /home/runner/work/fusionModel/fusionModel/docs/reference/fusion_model.fsn 
#> Total processing time: 7.05 secs