Read fusion output from disk

Read fusion output that was written directly to disk via fuse.

Usage

read_fsd(fsd, cores = data.table::getDTthreads())

Arguments

fsd

Character. File path ending in .fsd as produced by call to fuse.

cores

Integer. Number of cores used by fread.

Value

A data.table with integer column "M" indicating the implicate assignment of each observation. Note that the ordering of recipient observations is consistent within implicates, so do not change the row order if using with analyze or validate.

Details

The file written by fuse includes specially formatted metadata in the column names that read_fsd uses to construct the appropriate output values and classes. Reading fsd with a different file reader function will not give correct results.

Examples

# Build a fusion model using RECS microdata
# Note that "fusion_model.fsn" will be written to working directory
?recs
fusion.vars <- c("electricity", "natural_gas", "aircon")
predictor.vars <- names(recs)[2:12]
fsn.path <- train(data = recs, y = fusion.vars, x = predictor.vars)

# Write fusion output directly to disk
# Note that "results.fsd" will be written to working directory
recipient <- recs[predictor.vars]
sim <- fuse(data = recipient, fsn = fsn.path, M = 5, csv = "results.fsd")

# Read the fusion output saved to disk
sim <- read_fsd(sim)
head(sim)