Train a fusion model

Train a fusion model on "donor" data using sequential LightGBM models to model the conditional distributions. The resulting fusion model (.fsn file) can be used with fuse to simulate outcomes for a "recipient" dataset.

Usage

train(
  data,
  y,
  x,
  fsn = "fusion_model.fsn",
  weight = NULL,
  nfolds = 5,
  nquantiles = 3,
  nclusters = 2000,
  krange = c(10, 500),
  hyper = NULL,
  fork = FALSE,
  cores = 1
)

Arguments

data

Data frame. Donor dataset. Categorical variables must be factors and ordered whenever possible.

y

Character or list. Variables in data to eventually fuse to a recipient dataset. Variables are fused in the order provided. If y is a list, each entry is a character vector possibly indicating multiple variables to fuse as a block.

x

Character or list. Predictor variables in data common to donor and eventual recipient. If a list, each slot specifies the x predictors to use for each y.

fsn

Character. File path where fusion model will be saved. Must use .fsn suffix.

weight

Character. Name of the observation weights column in data. If NULL (default), uniform weights are assumed.

nfolds

Numeric. Number of cross-validation folds used for LightGBM model training. Or, if nfolds < 1, the fraction of observations to use for training set; remainder used for validation (faster than cross-validation).

nquantiles

Numeric. Number of quantile models to train for continuous y variables (along with the conditional mean). nquantiles evenly-distributed percentiles are used. The default nquantiles = 3 is usually sufficient and yields quantile models for the 16.6th, 50th, and 83.3rd percentiles.

nclusters

Numeric. Maximum number of k-means clusters to use. Higher is better but at computational cost. nclusters = 0 or nclusters = Inf turn off clustering.

krange

Numeric. Minimum and maximum number of nearest neighbors to use for construction of continuous conditional distributions. Higher max(krange) is better but at computational cost.

hyper

List. LightGBM hyperparameters to be used during model training. If NULL, default values are used. See Details and Examples.

fork

Logical. Should parallel processing via forking be used, if possible? See Details.

cores

Integer. Number of physical CPU cores used for parallel computation. When fork = FALSE or on Windows platform (since forking is not possible), the fusion variables/blocks are processed serially but LightGBM uses cores for internal multithreading via OpenMP. On a Unix system, if fork = TRUE, cores > 1, and cores <= length(y) then the fusion variables/blocks are processed in parallel via mclapply.

Value

A fusion model object (.fsn) is saved to fsn.

Details

When y is a list, each slot indicates either a single variable or, alternatively, multiple variables to fuse as a block. Variables within a block are sampled jointly from the original donor data during fusion. See Examples.

y variables that exhibit no variance or continuous y variables with less than 10 * nfolds non-zero observations (minimum required for cross-validation) are automatically removed with a warning.

The fusion model written to fsn is a zipped archive created by zip containing models and data required by fuse.

The hyper argument can be used to specify the LightGBM hyperparameter values over which to perform a "grid search" during model training. See here for the full list of parameters. For each combination of hyperparameters, nfolds cross-validation is performed using lgb.cv with an early stopping condition. The parameter combination with the lowest loss function value is used to fit the final model via lgb.train. The more candidate parameter values specified in hyper, the longer the processing time. If hyper = NULL, a single set of parameters is used LightGBM default values. Typically, users will only have reason to specify the following parameters via hyper:

boosting

num_leaves

bagging_fraction

feature_fraction

max_depth

min_data_in_leaf

num_iterations

learning_rate

max_bin

min_data_in_bin

max_cat_threshold

Testing with small-to-medium size datasets suggests that forking is typically faster than OpenMP multithreading (the default). However, forking will sometimes "hang" (continue to run with no CPU usage or error message) if an OpenMP process has been previously used in the same session. The issue appears to be related to Intel's OpenMP implementation (see here). This can be triggered when other operations are called before train() that use data.table or fst in multithread mode. If you experience hanged forking, try calling data.table::setDTthreads(1) and fst::threads_fst(1) immediately after library(fusionModel) in a new session.

Examples

# Build a fusion model using RECS microdata
# Note that "fusion_model.fsn" will be written to working directory
?recs
fusion.vars <- c("electricity", "natural_gas", "aircon")
predictor.vars <- names(recs)[2:12]
fsn.path <- train(data = recs, y = fusion.vars, x = predictor.vars)
#> 3 fusion variables
#> 11 initial predictor variables
#> 5686 observations
#> Using all available predictors for each fusion variable
#> -- Training step 1 of 3: electricity
#>   -- R-squared of cluster means: 0.97 
#>   -- Number of neighbors in each cluster:
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>    10.0    72.0   147.0   194.8   301.0   499.0 
#> -- Training step 2 of 3: natural_gas
#>   -- R-squared of cluster means: 0.977 
#>   -- Number of neighbors in each cluster:
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>    10.0    66.0   147.0   195.3   322.2   497.0 
#> -- Training step 3 of 3: aircon
#> Fusion model saved to:
#>  /home/runner/work/fusionModel/fusionModel/docs/reference/fusion_model.fsn 
#> Total processing time: 6.37 secs 

# When 'y' is a list, it can specify variables to fuse as a block
fusion.vars <- list("electricity", "natural_gas", c("heating_share", "cooling_share", "other_share"))
fusion.vars
#> [[1]]
#> [1] "electricity"
#> 
#> [[2]]
#> [1] "natural_gas"
#> 
#> [[3]]
#> [1] "heating_share" "cooling_share" "other_share"  
#> 
train(data = recs, y = fusion.vars, x = predictor.vars)
#> 5 fusion variables
#> 11 initial predictor variables
#> 5686 observations
#> Using all available predictors for each fusion variable
#> -- Training step 1 of 3: electricity
#>   -- R-squared of cluster means: 0.974 
#>   -- Number of neighbors in each cluster:
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>    10.0    69.0   152.0   199.4   316.0   499.0 
#> -- Training step 2 of 3: natural_gas
#>   -- R-squared of cluster means: 0.975 
#>   -- Number of neighbors in each cluster:
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>    10.0    69.0   152.5   193.9   308.2   499.0 
#> -- Training step 3 of 3: heating_share, cooling_share, other_share
#> Fusion model saved to:
#>  /home/runner/work/fusionModel/fusionModel/docs/reference/fusion_model.fsn 
#> Total processing time: 14.1 secs 

# When 'x' is a list, it specifies which predictor variables to use for each 'y'
xlist <- list(predictor.vars[1:4], predictor.vars[2:8], predictor.vars)
xlist
#> [[1]]
#> [1] "income"    "age"       "race"      "education"
#> 
#> [[2]]
#> [1] "age"         "race"        "education"   "employment"  "hh_size"    
#> [6] "division"    "urban_rural"
#> 
#> [[3]]
#>  [1] "income"      "age"         "race"        "education"   "employment" 
#>  [6] "hh_size"     "division"    "urban_rural" "climate"     "renter"     
#> [11] "home_type"  
#> 
train(data = recs, y = fusion.vars, x = xlist)
#> 5 fusion variables
#> 11 initial predictor variables
#> 5686 observations
#> Using specified set of predictors for each fusion variable
#> -- Training step 1 of 3: electricity
#>   -- R-squared of cluster means: 0.918 
#>   -- Number of neighbors in each cluster:
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>    10.0   107.0   222.0   240.6   379.0   499.0 
#> -- Training step 2 of 3: natural_gas
#>   -- R-squared of cluster means: 0.957 
#>   -- Number of neighbors in each cluster:
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>    10.0    64.0   172.0   212.6   357.8   499.0 
#> -- Training step 3 of 3: heating_share, cooling_share, other_share
#> Fusion model saved to:
#>  /home/runner/work/fusionModel/fusionModel/docs/reference/fusion_model.fsn 
#> Total processing time: 13.1 secs 

# Specify a single set of LightGBM hyperparameters
train(data = recs, y = fusion.vars, x = predictor.vars,
      hyper = list(boosting = "goss",
                   feature_fraction = 0.8,
                   num_iterations = 300
      ))
#> 5 fusion variables
#> 11 initial predictor variables
#> 5686 observations
#> Using all available predictors for each fusion variable
#> -- Training step 1 of 3: electricity
#>   -- R-squared of cluster means: 0.968 
#>   -- Number of neighbors in each cluster:
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>      10      68     151     194     301     499 
#> -- Training step 2 of 3: natural_gas
#>   -- R-squared of cluster means: 0.963 
#>   -- Number of neighbors in each cluster:
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>    10.0    62.0   149.0   189.5   297.0   497.0 
#> -- Training step 3 of 3: heating_share, cooling_share, other_share
#> Fusion model saved to:
#>  /home/runner/work/fusionModel/fusionModel/docs/reference/fusion_model.fsn 
#> Total processing time: 16.7 secs 

# Specify a range of LightGBM hyperparameters to search over
# This takes longer, because there are more models to test
train(data = recs, y = fusion.vars, x = predictor.vars,
      hyper = list(num_leaves = c(10, 30),
                   feature_fraction = c(0.7, 0.9),
                   num_iterations = 50
      ))
#> 5 fusion variables
#> 11 initial predictor variables
#> 5686 observations
#> Using all available predictors for each fusion variable
#> -- Training step 1 of 3: electricity
#>   -- R-squared of cluster means: 0.975 
#>   -- Number of neighbors in each cluster:
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>      10      89     185     216     339     498 
#> -- Training step 2 of 3: natural_gas
#>   -- R-squared of cluster means: 0.984 
#>   -- Number of neighbors in each cluster:
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>    10.0    62.0   141.0   183.1   280.5   498.0 
#> -- Training step 3 of 3: heating_share, cooling_share, other_share
#> Fusion model saved to:
#>  /home/runner/work/fusionModel/fusionModel/docs/reference/fusion_model.fsn 
#> Total processing time: 32.2 secs